day5/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
fastapi==0.115.12
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
